+++ template_files/api/base.rb
@@ -1,11 +1,9 @@
 module API
   class Base < Grape::API
 
-    include API::Helpers::CacheDSL
+    helpers API::Helpers::PaginateHelpers
 
-    helpers API::Helpers::CacheHelpers
     helpers API::Helpers::ApplicationHelpers
-    helpers API::Helpers::AuthHelpers
 
     prefix Application::Config.api_prefix_path if Application::Config.enabled?(:prefix_api_path)
