+++ template_files/api/helpers/cache/cache_helpers.rb
