+++ template_files/api/v1/routes/users_auth_social.rb
@@ -3,7 +3,7 @@ module API
     module Routes
       class UsersAuthSocial < API::V1::Base
 
-        helpers API::Helpers::V1::UsersHelpers
+        helpers API::Helpers::V1::UserAuthHelpers
 
         namespace :users do
           namespace :oauth do
