+++ template_files/api/v1/routes/users_me_cacheable.rb
@@ -3,7 +3,7 @@ module API
     module Routes
       class UsersMeCacheable < API::V1::Base
 
-        helpers API::Helpers::V1::UsersHelpers
+        helpers API::Helpers::V1::UserAuthHelpers
 
         before do
           authenticate_user
