+++ template_files/api/v1/routes/users_me_notifications.rb
@@ -11,19 +11,6 @@ module API
 
         namespace :users do
           namespace :me do
-            desc 'Create a new device in Parse'
-            params do
-              requires :device, type: Hash do
-                requires :token, type: String
-                requires :platform, type: String, values: ::UserDevice::VALID_PLATFORMS.values
-              end
-            end
-            post :devices do
-              service = execute_service('Parse::DeviceCreateService', current_user, params.to_h)
-
-              simple_response_for_service(service)
-            end
-
             desc 'Flag notification as read'
             put '/notifications/:id' do
               notification = current_user.notifications.unread.find_by(id: params[:id])
@@ -43,9 +30,8 @@ module API
               desc 'Get paginated notifications for current user'
               get :notifications do
                 response_for_paginated_endpoint 'current_user.paginated_notifications' do
-                  notifications = paginated_notifications_for_user(current_user)
-
-                  paginated_serialized_notifications(notifications, serializer: :current_user_notification).as_json
+                  options = { serializer: :current_user_notification }
+                  paginated_serialized_notifications(current_user.notifications, options).as_json
                 end
               end
             end # paginated_endpoint
